test: show DL button in testings Web UI
