+++ pkg/api/api/dev/test/works/index.ts
@@ -1,3 +1,1 @@
-import { defineController } from './$relay'
-
-export default defineController(() => ({}))
+export type Methods = Record<string, never>
